Fixing App Dynamics Agent to 4.3 chain
Locking App Dyanmics Agent to latest version in 4.3.*.*. This is required to have version of the agent be at or lower than the App Dynamics Controller version. In the case of dedicated SaaS the controller is not always upto date with the latest agent. (Presently, the agent is 4.4.??, but the controller is still 4.3.4.1.